requirements.txt
ccxt>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
aiohttp>=3.9.0
